[REF] Char, Inline: move format cache to char
