+++ packages/plugin-inline/Inline.ts
@@ -3,7 +3,6 @@ import { RangeParams } from '../core/src/CorePlugin';
 import { Format } from './Format';
 import { InlineNode } from './InlineNode';
 import { Constructor } from '../utils/src/utils';
-import { VNode } from '../core/src/VNodes/VNode';
 
 export interface FormatParams extends RangeParams {
     FormatClass: Constructor<Format>;
@@ -15,15 +14,6 @@ export class Inline extends JWPlugin {
             handler: this.toggleFormat.bind(this),
         },
     };
-    commandHooks = {
-        setSelection: this.resetFormatCache.bind(this),
-    };
-    /**
-     * When apply format on a collapsed range, cache the calculation of the
-     * format the following property. This value is reset each time the range
-     * change in a document.
-     */
-    formatCache: Format[] = null;
 
     //--------------------------------------------------------------------------
     // Public
@@ -38,68 +28,33 @@ export class Inline extends JWPlugin {
     toggleFormat(params: FormatParams): void {
         const range = params.range || this.editor.vDocument.selection.range;
         const FormatClass = params.FormatClass;
-        if (range.isCollapsed()) {
-            if (!this.formatCache) {
-                this.formatCache = this.getCurrentFormats();
-            }
-            const index = this.formatCache.findIndex(f => f instanceof FormatClass);
-            if (index !== -1) {
-                this.formatCache.splice(index, 1);
-            } else {
-                this.formatCache.push(new FormatClass());
-            }
-        } else {
-            const selectedInlines = range.selectedNodes(InlineNode);
 
-            // If every char in the range has the format `FormatClass`, remove
-            // the format for all of them.
-            const allHaveFormat = selectedInlines.every(inline => {
-                return !!inline.formats.find(f => f instanceof FormatClass);
-            });
-            if (allHaveFormat) {
-                for (const inline of selectedInlines) {
-                    const index = inline.formats.findIndex(f => f instanceof FormatClass);
-                    const matchingFormat = inline.formats[index];
-                    for (const key of Object.keys(matchingFormat.attributes)) {
-                        inline.attributes[key] = matchingFormat.attributes[key];
-                    }
-                    inline.formats.splice(index, 1);
+        if (range.isCollapsed()) return;
+
+        const selectedInlines = range.selectedNodes(InlineNode);
+
+        // If every char in the range has the format `FormatClass`, remove
+        // the format for all of them.
+        const allHaveFormat = selectedInlines.every(inline => {
+            return !!inline.formats.find(f => f instanceof FormatClass);
+        });
+        if (allHaveFormat) {
+            for (const inline of selectedInlines) {
+                const index = inline.formats.findIndex(f => f instanceof FormatClass);
+                const matchingFormat = inline.formats[index];
+                for (const key of Object.keys(matchingFormat.attributes)) {
+                    inline.attributes[key] = matchingFormat.attributes[key];
                 }
-                // If there is at least one char in the range without the format
-                // `FormatClass`, set the format for all nodes.
-            } else {
-                for (const inline of selectedInlines) {
-                    if (!inline.formats.find(f => f instanceof FormatClass)) {
-                        new FormatClass().applyTo(inline);
-                    }
+                inline.formats.splice(index, 1);
+            }
+            // If there is at least one char in the range without the format
+            // `FormatClass`, set the format for all nodes.
+        } else {
+            for (const inline of selectedInlines) {
+                if (!inline.formats.find(f => f instanceof FormatClass)) {
+                    new FormatClass().applyTo(inline);
                 }
             }
         }
-    }
-    /**
-     * Get the format for the next insertion.
-     */
-    getCurrentFormats(range = this.editor.vDocument.selection.range): Format[] {
-        if (this.formatCache) {
-            return this.formatCache;
-        }
-
-        let inlineToCopyFormat: VNode;
-        if (range.isCollapsed()) {
-            inlineToCopyFormat = range.start.previousSibling() || range.start.nextSibling();
-        } else {
-            inlineToCopyFormat = range.start.nextSibling();
-        }
-        if (inlineToCopyFormat && inlineToCopyFormat.is(InlineNode)) {
-            return [...inlineToCopyFormat.formats];
-        }
-
-        return [];
-    }
-    /**
-     * Each time the selection changes, we reset its format.
-     */
-    resetFormatCache(): void {
-        this.formatCache = null;
     }
 }
